refactor(frontend): Made bus details in drawer dynamic and created a separate component for it
